pyproject.toml
[tool.poetry]
name = "augmix-refactored"
version = "0.1.0"
description = "A refactored version of the original augmix project."
authors = ["Jan Philipp Schneider <jp-schneider@users.noreply.github.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "augmix_refactored"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
numpy = "^1.24.2"
Pillow = "^9.4.0"
torch = {version=">1.0", optional=true}
torchvision = {version=">0.14", optional=true}
typing-inspect = "^0.8.0"
simple-parsing = "^0.0.21.post1"
pyyaml = "^6.0"


[tool.poetry.dev-dependencies]
pytest = "7.1.2"
pylint = "^2.13.3"
pycodestyle = "^2.8.0"
autopep8 = "^1.6.0"

[tool.poetry.extras]
torch = ["torch", "torchvision"]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
